refactor: refactor the schedule module to avoid cyclic imports
